+++ client/src/App.jsx
@@ -34,7 +34,7 @@ export default function App() {
           <Route path="/leaderboard" element={<Leaderboard />} />
           <Route path="/profile" element={<Profile user={user} setUser={setUser} />} />
           <Route path="/register" element={<Register />} />
-          <Route path="/login" element={<Login user={user} setUser={setUser} />} />
+          <Route path="/login" element={<Login setUser={setUser} />} />
           <Route path="/success/:category" element={<SuccessfulSubmission />} />
           <Route path="/404" element={<PageNotFound />} />
           <Route path="*" element={<Navigate to="/404" replace />} />
